+++ src/Commands/Env/Update.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Sven\ForgeCLI\Commands\Env;
+
+use Sven\ForgeCLI\Commands\BaseCommand;
+use Symfony\Component\Console\Input\InputOption;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
+use Symfony\Component\Console\Output\OutputInterface;
+
+class Update extends BaseCommand
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function configure()
+    {
+        $this->setName('update:env')
+             ->addArgument('server', InputArgument::REQUIRED, 'The id of the server the site is on.')
+             ->addArgument('site', InputArgument::REQUIRED, 'The id of the site you want to update the .env file of.')
+             ->addOption('file', null, InputOption::VALUE_REQUIRED, 'The path to your new .env file.')
+             ->setDescription('Update the .env file of the given site.');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function perform(InputInterface $input, OutputInterface $output)
+    {
+        $this->forge->updateSiteEnvironmentFile(
+            $input->getArgument('server'), $input->getArgument('site'), $this->getFileContent($input, 'file')
+        );
+    }
+}